manual: Update stale references to magit-insert-submodules
The manual contains two references to magit-insert-submodules, which
was renamed to magit-insert-modules-overview.  One of the stale
references is a suggestion to add magit-insert-submodules to
magit-status-sections-hook.  Instead of substituting in the new name,
recommend the magit-insert-modules wrapper (introduced in the same
series as the rename).

Just delete the other reference, a function entry, because it is
already covered by an entry for magit-insert-modules-overview in
"Status Module Sections".

Closes #4162.