magit-section-maybe-update-visibility-indicator: Bugfix
We used `face-foreground' because before that we used a single
hard-coded color string.  When we stopped doing that, then we
should also have stopped using a single color and switched to
a face.

Because we didn't, the use of `text-scale-adjust' led to the
use of `default's background color as fallback in the fringe.

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>