magit-completing-read-multiple: Override helm-crm-default-separator
As of Helm v3.6.2 (specifically commit 21261899, 2020-05-03), Helm
appends the separator to the completing-read-multiple input after the
caller completes an item.  That behavior works fine for the default
completing-read-multiple and magit-completing-read-multiple*, where
the input is split on the separator and thus the particular value of
the separator isn't of interest.  The caller is saved from typing the
separator, and the trailing separator is dropped in the split.

On the other hand, magit-completing-read-multiple does not split the
result, and the presence of a separator as well as its particular
value is usually meaningful (e.g., ".." versus "..."  in ranges).
Automatically appending an inferred separator creates unintended and
often invalid results.

Bind the recently added helm-crm-default-separator (0e77797e,
81dbe5e9) to nil to prevent the separator from being appended,
restoring the previous behavior.

Ref: https://github.com/emacs-helm/helm/issues/2353