magit-{tag-delete,stash-drop}: Supports Helm's marked candidates
Helm's completing-read variant recently gained support for marking
multiple items, which are returned as a list.  To enable this
functionality, Lisp-callers must let-bind helm-comp-read-use-marked
around their completing-read calls.

There are a few sites in Magit that this would work with as is.
Adjust the two that seem useful, leaving magit-revert-read-args,
magit--cherry-move-read-args, and magit-cherry-pick-read-args
untouched.

Modified-by: Kyle Meyer <kyle@kyleam.com>
  * Wrote commit message.
  * Added magit-tag-delete support.

Closes #4185.