Enable global-magit-file-mode in autoloads file
This ensures that magit-file-mode-map is active as soon as the user
visits a file under version control, without them having to invoke
magit-status or (require 'magit) first.

global-magit-file-mode--activated is here to make sure Emacs does not
enable the mode twice - once when loading magit-autoloads.el, and
again when loading magit-file-modes.el.

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>