Add three global key bindings by default
For the longest time we have tried hard to avoid doing this but
it is just not worth the effort and leads to an inferior result.

Now we add three bindings by default.

  ;;;###autoload
  (when magit-define-global-key-bindings
    (let ((map (current-global-map)))
      (define-key map (kbd "C-x g")   'magit-status)
      (define-key map (kbd "C-x M-g") 'magit-dispatch)
      (define-key map (kbd "C-c M-g") 'magit-file-dispatch)))

As you can see is easy to prevent these bindings from being added
and I hope having to do so won't upset anyone too much.  While we
did not add global bindings we already added these bindings in all
file-visiting buffers and we did not get any complaints about that.

This completely replaces `magit-file-mode' and the global variant,
which we define as an alias of the new variable.

See #4237.