magit-insert-section: Don't require magit--section-type-alist
When TYPE is a class, then do not require there to be a corresponding
entry in `magit--section-type-alist'.  Such entries are only required
for legacy section types that are symbols that do not match the name
of the corresponding section classes.  Existing third-party section
types likely have no corresponding class yet (but the names should be
prefixed).

Now third-party section types can be defined properly, using a class,
without the author having to somehow figure out that they also have
to add an entry to `magit--section-type-alist'.

Closes #4243.