git-commit-setup-font-lock: Pacify debug-on-signal
We have to be able to enable `debug-on-signal', so we cannot have
this function needlessly enter debugger due to an expected signal.

`process-lines' signals an error if the process exits with a non-zero
status.  That is expected here and we used `ignore-errors' to deal
with that case, but `debug-on-signal' overrides that.