magit-section--open-temporarily: Bugfix
A section may be invisible because the body of its parent is hidden,
in which case that section does not necessarily have a body of its
own, which also implies that the `content' slot is nil.

In that case the parent has to be expanded but the section itself does
not and attempting to do so resulted in an error due to an attempt to
perform an arithmetic operation on `content', which is nil.