magit-log-bisect-log-re: Match arbitrary terms
Since Git v2.7.0, bisect has allowed terms other than "bad" and
"good".  Even if we don't support performing a bisect with alternate
terms from Magit, we still need to be able to parse the bisect log to
avoid failures when a user starts a bisect with alternative from
outside Magit and then visits the repository in Magit.

Extend the bisect log regular expression to support arbitrary terms
(which are limited to valid ref name characters).