bisect: Support specifying alternate terms
Alternate terms are useful for bisecting a change that conceptually
doesn't fit the default "bad" and "good" terms.  Add infix and suffix
arguments to the magit-bisect transient to expose this functionality,
but hide them away by default to avoid complicating the transient for
those that don't have a use for the more general interface.

Note that unlike "good" and "bad", which are fontified with different
faces (magit-bisect-good vs magit-bisect-bad), both alternate terms
are fontified with the same face, magit-refname.  While it'd perhaps
be helpful to distinguish between the refs for each term, it doesn't
seem worth the trouble to address.