git-commit-prev-message: Don't reorder message ring
The previous implementation messed up the order of the ring.
The only change we intended to make was to add the newly written
message (if any) at the top of the ring.

Therefore using `git-commit-save-message' was inappropriate and we
now use a copy of `log-edit-previous-comment', which we fix to not
lose the initial message and to preserve the comment at the end of
the message buffer.

Closes #4283.