magit-diff: Implement jumping to {orig-worktree} buffers
When point is on a removed line in the "unsaved changes" diff, make
RET (magit-visit-thing) visit a buffer representing the "old" side of
the diff, i.e., a buffer tracking the file as it is on disk right now.

Use the {orig-worktree} pseudo-revision added in the previous commit
for that purpose.