Set the diffstat width to window's width with new customization
Before this patch the filename + graph lines of diffstat are limited to 78 characters by
git (according with the git documentation this is 80 characters, but the limit is actually 78 in my
tests). This causes part of the directory structure and/or filename to be abbreviated to 78
characters even if the buffer has enough space to display the full directory structure and filename.

This patch set the default width of diffstat to the same as the window's width and add 3 new custom
variables to customize this behavior in the same way as available when using git in a terminal.

This patch affects the magit-diff, magit-revision and the magit-log buffers.