magit-rebase-remove-commit: override core.commentChar
When "magit-rebase-remove-commit" (r k) is called, it replaces the line with
"noop" followed by a commented-out version of the original "pick" command on a
separate line. However, the comment character is hardcoded to "#", which does
not work as expected when the "core.commentChar" option is set to something
other than "#" in .gitconfig files.

This change overrides the value of "core.commentChar" when
"magit-rebase-interactive-1" is called within "magit-rebase-remove-commit", so
that the rebase operation can properly delete the commit, and further ensures
that it set when "magit-log-select" called in "magit-rebase-interactive-1".