No longer autoload global-git-commit-mode
The autoloads triggered full loading as soon as the user created
any buffer.  The autoloads had to do that to be of any use, which
means that they had nearly the same effect as loading the library
explicitly, which unfortunately is very slow.

Now that the autoloads are gone, users can actually choose whether
they want to pay that price.  Most users likely won't notice that
anything has changed as this only matter when one sometimes commits
from the command-line without loading `magit' first.

Closes #4293.