remove git-commit-major-mode entry from .dir-locals.el
This variable is not *safe* varibale, warning window is noisy for
new contributors.  Answering y to warnings is a bad habit to get
beginners into.  Think about what it means for .dir-locals to
show a warning for setting an unsafe variable, and if we need
this, we should do it in dir-locals-2.el.