magit-extras: Update project.el integration
The format of project-switch-commands changed.  The car is now the
command, and its key binding is looked up in project-prefix-map:

Remove the duplication from project-switch-commands's config
51698f77dd 2020-12-13 22:50:55 +0200
https://git.savannah.gnu.org/cgit/emacs.git/commit/?id=51698f77dd6356049fcacdb01ebe80cfe4c67272

This change is included in project.el version 0.5.4 of 2021-03-09,
which also lowers the minimum required Emacs version from 26.3 to
26.1.

Consistently calling magit-project-status instead of magit-status
has the benefit of going through project-remember-project, which
marks the selected project as the most recently used, for the
benefit of completion.