Propagate git failures that are due to corrupt configuration
magit-toplevel uses a failing `git rev-parse --show-toplevel' to
decide that it's not in a working tree.  In case of corrupt Git
configuration, it will incorrectly conclude that it's not in a
repository and won't propagate the error to the user.  These two
failures can't be distinguished by exit code.  We can check for a
corrupt configuration with a follow-up git-config call, but it's
probably not worth doing that within the heavily used magit-toplevel.

Instead do it as part of the existing
magit--not-inside-repository-error handling as well as when an
interactive call to magit-status doesn't find a repository.  This
covers magit-status, magit-status-setup-buffer, and
magit-get-mode-buffer, so it should ensure that the error is relayed
to the caller.

Closes #4337.