magit-refresh-buffer: Fix window handling
Previously we assumed that the `window-start' is associated with a
window AND a buffer, but this is wrong.  Each window only has one
`window-start'.

When the buffer was not displayed in any window, then we made the
buffer current in the selected window (using `with-current-buffer')
and then called `magit-section-goto-successor', which is fine except
that that may end up running `magit-section-movement-hook' and one of
the functions on that hook (`magit-hunk-set-window-start') may call
`set-window-start'.

(This only happened if current section was a hunk and the user edited
the respective file in a way that caused that hunk to disappear, which
turned into a long-lived heisenbug.)

The `window-start' is not actually adjusted until Emacs redisplays at
which point the `with-current-buffer' is no longer in effect, so the
window start adjustment affects whatever buffer is actually current in
the selected window, essentially causing point to jump to some random
location.

To avoid this, we suppress `magit-section-goto-successor-hook' when
the buffer is not current in any window.  We do not try to only
suppress `magit-hunk-set-window-start', but suppress all hook
functions because the other functions are also undesirable when the
buffer is not current, just not as fatally.

Closes #4196.