magit-bisect-run: Don't start bisect asynchronously
As of 56354898 (Run "git bisect" asynchronously, 2019-04-20),
magit-git-bisect runs `git bisect' calls asynchronously.  While
beneficial in most spots, this switch creates a race if
magit-bisect-run is called when a bisect hasn't been started.

In that case, magit-bisect-run calls magit-bisect-start (which calls
the asynchronous magit-git-bisect underneath) and then immediately
follows with a `git bisect run' call via magit-git-bisect.  This
second call depends on the first asynchronous call completing, so it
leads to a "Not bisecting" user error.  That's quickly buried when the
magit-bisect-start call finishes, and the end result looks like the
user just called magit-bisect-start.

Before 56354898, magit-git-bisect used the now removed
magit-run-git-with-logfile.  Inline that code for magit-bisect-run's
start call.