magit-diff-wash-signature: Collapse harmless output
If the signature is good or cannot be checked because the cert
is missing, then collapse the section body.  Otherwise continue
to show all off gpg's output.

If the signature cannot be verified because the cert is missing,
then that isn't worse than if the commit weren't signed at all,
so do not use a warning face.  Since [1: 3afd4a394] the styling
of `magit-signature-error' takes this into account but because
the color was chosen to work well when used to display the hash
in a log, it doesn't work well here.  We still want the "missing
cert" notice to stand out a bit but without being alarmist, so
we combine the `italic' and `bold' faces.

1: 2021-02-14 3afd4a394439e8e6005b4b921a070c976f3b8909
   magit-signature-error: Use a less alarmist color