magit-stash-drop: Use "git stash drop"
Previously we called "git rev-parse --verify refs/stash@{0}" and if
that didn't return a revision then we knew that no stash remains and
so we also deleted the "refs/stash" reference itself.  We cannot do
that anymore because a special case was added in 6436a20284f33d421,
which makes that command return a revision even when the reflog is
empty.  Now we use "git stash drop" directly.  It uses the same logic
to determine whether any stash remains before deleting the ref, but
it does so using a lower-level C api, which is not affected by this
backward incompatible change in git.  We did not use "git stash drop"
when we first implemented `magit-stash-drop' because "git-stash" was
still a shell-script and just copying the relevant bits made more
sense.

Closes #4452.