magit-diff-wash-diff: Don't match renames in 'diff --git' line
In the

  diff --git FILE1 FILE2

header, FILE1 and FILE2 are the same in all cases except for renames
and copies.  magit-diff-wash-diff's regexp for matching this line used
to have a group that matched FILE1 and FILE2 if they were the same;
otherwise, the regexp would match but without the file group.  The
copied or renamed files would then be set later via the subsequent
copy/rename header lines.

d477163d (magit-diff-wash-diff: Add kludge for inconsistent git
output, 2021-07-12) extended magit-diff-wash-diff's 'diff --git'
regexp to match FILE1 and FILE2 even if the files differ.  This works
in the simple case, but for more complicated cases, in particular
those involving spaces, the header line isn't parsed reliably.  Here
are some examples of problems this leads to:

  * Deleting a file "foo bar" results in an incorrect section

      deleted    foo bar -> bar foo bar

  * Adding a new file "foo bar" results in an incorrect section

      new file   foo -> foo bar

  * Adding a new file with a space as well as a character that causes
    git to quote the file name leads to a parse error because
    magit-decode-git-path is called with a file name that has an
    opening quote but not a closing one.

    This is the cause of the current magit-status:file-sections test
    failure.

Fix these cases by restoring the same name restriction to the file
name group capture, letting the file names for renames and copies be
extracted from the subsequent copy/rename lines that can be reliably
parsed.