magit-wip-log-get-tips: Fix retrieval of previous tips
magit-wip-update-wipref switched the reflog message from "restart
autosaving ..." to "start autosaving ..." in 40c06c51 (wipref: Support
automatic merging of real branch into wip ref, 2018-08-02).
magit-wip-log-get-tips searches for this message in order to find
previous tips of the wip refs, but its regular expression wasn't
updated with that commit.

Teach magit-wip-log-get-tips to look for the new reflog message in
addition to the old one.  With Git's default gc settings, it's pretty
unlikely to encounter the old message at this point, but a user may
have set gc.reflogExpire to "never".