magit-process-kill-on-abort: Evaluate kbd during macro expansion
413a98f7 (Use kbd function more when defining key bindings,
2021-08-13) replaced a "\C-g" in the magit-process-kill-on-abort macro
with (kbd "C-g").  This causes the magit-process:password-prompt test
to sit waiting for input at

  Password for 'www.host.com':

That's with Emacs 27.2 on my local machine.  The test is also failing
in the GitHub Actions build.

I don't understand why the kbd switch bothers that test, and, as far
as I can tell, magit-process-kill-on-abort is functioning as intended
outside of the tests.  But evaluating the kbd form during macro
expansion fixes the test failure on my machine, so switch to that.