Fix CI by relying less on Nix
The previous implementation kept breaking and we got no help fixing
that.  More or less go back to the previous implementation, which we
know how to keep running without any outside help.

Previously testing using libgit was disabled because it was broken,
while now there simply isn't any code to attempt to test that.