git-commit-search-message-{backward,forward}: Don't use format-prompt
format-prompt isn't available until (what will be) Emacs 28.1.  Use a
plain format call instead.

The new behavior is different in that 1) it doesn't respect
minibuffer-default-prompt-format and 2) it shows "(default )" for an
empty string default rather than hiding the default entirely.  The
first seems an acceptable trade for compatibility while avoiding a
wrapper/kludge.  And the second seems fine in general given that the
empty string is in fact the default, so it's worth telling the user
about the useless default value.