magit-current-blame-chunk: Don't try to operate in Dired buffers
magit-current-blame-chunk passes the result of
magit-file-relative-name to git-blame.  For a Dired buffer, this value
is default-directory, and passing a directory to git-blame will lead
to an error.  This error can show up when various commands unrelated
to magit-blame are called from a Dired buffer because
magit-current-blame-chunk is called by magit-commit-at-point and
magit-branch-or-commit-at-point.

Don't try to blame the return value of magit-file-relative-name when
in a Dired buffer, letting the "no file" case be handled according
NOERROR.

Fixes #4560.