magit--shell-command: Use magit--with-connection-local-variables
Use magit--with-connection-local-variables macro to make sure that
`shell-file-name` and `shell-command-switch` have values appropriate
for the machine we are trying to run the command on.

Consider the scenario where we are using Emacs + Magit running on a
Windows machine while manipulating a remote repo on a Linux machine
with Tramp + plink. Without `magit--with-connection-local-variables`
we'll have `shell-file-name` pointing to `cmdproxy.exe` which, for
obvious reasons, won't work on the remote Linux box.