magit-sparse-checkout: Add --sparse-index
If `git sparse-checkout init' is called with --sparse-index, the index
is written with sparse directory entries.

Note that the user can also get this behavior by configuring
index.sparse to true, in which case --no-sparse-index would be the
useful argument to have in the transient so that the caller could
override their configured value for a particular call.

However, leave off --no-sparse-index, at least for now.  There are
other examples of --x/--no-x options along with an associated config
knob (e.g., --recurse-submodules in magit-branch) that we don't bother
supporting (in terms of having both options available in the transient
or considering the configured value).