magit-clone-sparse: New command
When the goal is to do a sparse checkout after cloning a repo, it's
possible to avoid checking out undesired files on the initial clone:

  $ git clone --no-checkout ...
  $ git sparse-checkout init --cone
  $ git checkout $BRANCH

Doing that probably isn't very common, so it's tempting to add
--no-checkout to the magit-clone transient and then let users handle
the sequence themselves.  However

  * it seems unlikely that many people would remember that sequence

  * those steps don't work well in Magit because, after the `clone
    --no-checkout` step, refreshing the status buffer will get hung up
    on all the "deleted" files.  That's true even with a repo of
    normal size, but it seems reasonable to expect larger target repos
    in the sparse checkout context.

Add a dedicated command that hooks into magit-clone-internal to enable
a sparse checkout before a refresh is attempted.