magit--with-connection-local-variables: Silence byte-compiler
With Emacs versions before 27.1, spots that use the recently added
magit--with-connection-local-variables compatibility kludge trigger
byte-compiler warnings about tramp-dissect-file-name and
tramp-get-method-parameter being unknown functions.  Given that this
code is executed only when `file-remote-p' returns non-nil, I suspect
that in practice it is unreachable without Tramp loaded, but load
Tramp explicitly to be sure.