magit-fetch-modules: Remove --all argument
The intention was to fetch --all remotes of all submodules, which may
or may not work.  It does cause all remotes of the super-repository to
be fetched, which wasn't the goal but which is perfectly fine.

However for each remote of the super-repository that if fetches, it
additionally goes through all the submodules and fetches each one of
them again.  We definitely don't want to needlessly fetch each
submodule N times, where N is the number of remotes of the
super-repository.