git-commit: Use magit-git-executable function if available
Three spots in git-commit.el hard code "git" as the git executable.
Two of these spots are in git-commit-self-ident.  If the user doesn't
have git in $PATH, these spots aren't much of a problem because 1)
these calls are wrapped with ignore-errors and have fallback values
and 2) their only callers are some of the git-commit-* "insert header"
commands.

The "git" in git-commit-setup-font-lock, on the other hand, can cause
committing to fail if git isn't in $PATH:

  1 git … commit --
  hint: Waiting for your editor to close the file...
  Waiting for Emacs...
  *ERROR*: Searching for program: No such file or directory, git
  [...]

A Magit user without git in $PATH must configure the
magit-git-executable variable, so teach git-commit.el to try to get
the executable path from the magit-git-executable function.  Using the
function means that the appropriate path is also used for Tramp
buffers.

Fixes magit/with-editor#106.