Refactor git version asserts
`magit-git-version' now displays an error message if Git cannot be
found, if it exits with a non-zero status, or the output does not have
the expected format.

The new function `magit-git-version-assert' uses the former function
and so displays a message for the same issues.  Additionally it checks
if Git is recent enough and if not it displays an error message about
that.

`magit--tramp-asserts' is no longer needed because the former function
now does everything it used to do.  Likewise `magit-startup-assert'
now longer has to display warnings about older Gits itself anymore.

`magit--host-git-version-cache' replaces
`magit--remotes-using-recent-git', which makes cache lookup more
expensive because it involves version comparison.  On the other hand
it allows comparing with different minimal versions, not just the
global minimum.  The use of `magit--with-refresh-cache' should also
help with the increased cost.

The displayed error messages are trimmed down a bit, removing some
snark and adding mentions of `magit-debug-git-executable'.