Remove redundant version/feature caches
Functions that compare the Git version on the current host with the
required version use `magit-git-version', which caches the used
version per host.

The result of comparing two versions is not cached, which increases
the cost slightly because they have to be converted to the internal
list format before comparison can happen.  `magit-git-version' uses
the refresh cache, making the effect negligible.