git-commit-setup-font-lock: Tighten git-config output processing
git-commit-setup-font-lock calls 'git config core.commentchar' via
call-process.  This call exits with 1 if the key isn't found.  On
success or failure, that normally shouldn't produce any stderr, and on
success the only stdout should be the core.commentchar.

But to be safe and reduce the chances of processing unintended output,
1) stop if there was a non-zero exit and 2) insert only standard
output into the buffer.