magit-insert-revision-message: Avoid hash section at message start
When magit-revision-use-hash-sections is non-nil,
magit-insert-revision-message searches the message for commits and
inserts (commit text) sections.  It starts at the beginning of the
(commit-message) section, which means it's possible to add a
(commit text) section that starts at the same spot as
(commit-message).

Having two sections at the same point probably breaks various things,
but at the very least it prevents magit-section-backward from moving
beyond the shared (commit-message)/(commit text) start and can send
magit-section--backward-find into an infinite loop.

Prevent this by starting the search for commits one character after
the start of the (commit-message) section.  This change means it's not
possible to convert the first word to a (commit text) section, but
it's doesn't seem worth doing anything more hacky or complicated to
support that.

Closes #4602.