Avoid updating section highlighting twice or prematurely
Some commands need to force updating of section highlighting, which
they used to do by calling `magit-section-highlight-update' with the
force argument, but because the highlighting is also updated on
`post-command-hook' that could result in it happening twice.

Now these commands can set `magit-section-highlight-force-update'
to cause the next update to be forced.