magit-diff-highlight-file: Always highlight module sections
Module sections are currently being treated as a special kind of file
sections, which may be a mistake, but determining that will require
additional research.  This commit fixes one issue that resulted from
this mismatch.

Unlike regular file sections, the children of modules are commits not
hunks, but the specialized code that delays highlighting diff-related
sections does not deal with commits.  For module sections the
highlighting was delayed, but then not performed at a later time.

Now we don't delay it anymore, which for these sections isn't
necessary anyway because doing it is as cheap as it is for other
non-diff sections, which also do not delay highlighting of children
that are hidden because the parent is collapsed.