Combine neighboring chunks about the same commit
When using, e.g., "--ignore-revs" together with "--incremental", it
can happen that neighboring chunks are actually about the same commit.

This only looks backward because that is all that was needed in the
test cases I used.  It is possible that we will have to extend this
to also look the other way.

Closes #4030.