magit-sequencer-continue: Abort for unmerged, not unstaged, content
magit-sequencer-continue signals a user error if there are any
unstaged changes.  However, that's stricter than the behavior of 'git
cherry-pick --continue' and 'git revert --continue', which are okay
with unstaged modifications in files without conflicts.

To support this behavior, one option would be to drop the guard
entirely and let 'git (cherry-pick|revert) --continue' do the error
reporting.  However, the reason the Lisp error was added in the first
place was in response users reporting these Git failures as a bug.

So instead signal a user-error if there are unmerged files.  It's
possible that this narrower condition lets through a few rare cases
that trigger a '(cherry-pick|revert) --continue' failure, but those
are hopefully uncommon enough that they won't prompt bug reports.

Closes #4615.