magit-ediff-read-files: Handle renames in one-file logs
When a log uses "--follow" to show the history for a single file,
including history from before it was renamed, then the user might
invoke Ediff on a range where the name of the modified file does
not match the name stored in `magit-buffer-log-files'.

Try to determine which of the files that are being modified in that
range is an early incarnation of the file for which the log is shown.

This fails if the differences are too large for `git-diff' to still
consider the files to be renamed.  In that case fall back to having
users select the file themselves.

Closes #4377.