Support revisions that match commit message
Git supports revisions that match a commit message, such as ":/fix
nasty bug".  See the description for ":/<text>" in gitrevisons(7).

Magit used to fail when encountering such revisions because it appends
"^{commit}" in several places to revisions in order to dereference
references.  This commit teaches it to recognize this special revision
syntax and just use such revision without dereferencing or verifying
them.

Mark the function as private because we might still have to tweak how
such special revisions are handled.  I.e., we might have to verify
such revisions (or not because at least callers already do that), and
if we do, we might have to return nil or raise an error (depending on
the caller).  For now at least, just do the easy thing, because this
syntax is only rarely used (never in Magit, once in `orgit'), so
putting a lot of effort into this isn't justified (yet?).

Co-authored-by: Jonas Bernoulli <jonas@bernoul.li>