magit-revert-rev-file-buffer: Use set-auto-mode
For that newly-opened buffer to have syntax coloring this function
briefly sets buffer-file-name and performs (normal-mode t).  This,
in turn, triggers related major-mode hooks, which at least in the
case of lsp causes issues.

An attempt was made to resolve this in f331092df but it didn't seem
to resolve the issue.

Because of how the buffers are named, set-auto-mode couldn't
work out which mode to use; so this has been swapped around.