magit-show-commit: Fix handling of current buffer's file
If magit-file-relative-name returns a file for the current buffer,
magit-show-commit expects the file's buffer to remain the current
buffer when calling save-buffer and line-number-at-pos.  However,
these calls are downstream of a magit-revision-setup-buffer call that
may have changed the current buffer to the revision buffer, depending
on the value of magit-display-buffer-noselect.

Due to the location of this line-number-at-pos call, invoking
magit-show-commit on a blame chunk does _not_ put point on the
relevant line while magit-diff-show-or-scroll-up does, because the
latter binds magit-display-buffer-noselect to t.

Move the save-buffer and line-number-at-pos calls before the
magit-revision-setup-buffer call.