magit-diff-wash-hunk: Handle one line edge case
Calling magit-show-commit on a blame chunk for a diff that has one
line in the post-image blob fails with a type error because
magit-diff--locate-hunk expects the diff range's length (N in "+M,N")
to be a number.  For a single line blob, this diff header codes this
as "+1", leading to magit-diff-wash-hunk setting to-range to (1).

Make magit-diff-wash-hunk catch this case and insert 1 for the length
so that downstream consumers like magit-diff--locate-hunk don't need
to worry about this edge case.