magit-commit-diff-1: Use different heuristics
Most importantly, do not show an alternative diff unless we are
confident that doing so is useful, or at least not highly confusing.
No longer derive the alternative diff from the arguments used for the
initial diff.  Ignore the last rebase command, because very often that
has nothing to do with the commit that is being created.

Closes #3246.