Default names to non-short format
The `refname:short` format is slow. On my repository (with 1000+ branches) it takes about 3.5seconds to show branches with `short` vs. about 0.2 seconds for the default format. Maybe this should be configurable somehow or depend on number of brances. Having `short` makes magit painful to use on repos with many branches. Tested on multiple versions of git and under windows and linux.