magit-commit-diff-1: Fix recent regression in --all handling
As of 2dfeaa68 (magit-commit-diff-1: Use different heuristics,
2022-06-07), magit-commit-diff-1 uses magit-anything-unstaged-p when
deciding what diff to show.  However, $GIT_DIR is the default
directory when magit-commit-diff-1 is called, so the 'git diff'
command used by magit-anything-unstaged-p fails with an exit code of
128, leading to magit-anything-unstaged-p returning nil even when
there are unstaged changes.

Switch to the top-level of the working tree so that
magit-anything-unstaged-p gives an accurate answer.

Closes #4725.