magit-push-implicitly--desc: Account for local upstream
The magit-push-implicitly--desc rewrite in the last commit doesn't
handle a local upstream such as

  branch.topic.remote=.
  branch.topic.merge=refs/heads/master

The problem is that it uses Git's fallback logic to determine the
remote when magit-get-remote returns nil, but magit-get-remote returns
nil when "branch.{branch}.remote" isn't set _or_ is set to ".".  For
the "." case, the description incorrectly reports that the push will
be to the fallback remote rather than the local repo.

Fix this issue by side stepping magit-get-remote and treating "." as
any other remote.  Showing "." as the remote seems accurate enough
given that the underlying push command will push to the local
repository, and it shows the "remote" as ".".

While touching this block of code, also pass the branch to
magit-get-push-remote so that it doesn't need to repeat the
magit-get-current-branch call itself.